// client/src/app/api/models/run-event-envelope.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RunEvent } from '../models/run-event';
import { RunKind } from '../models/run-kind';
export interface RunEventEnvelope {
  event?: RunEvent;
  isTerminal?: boolean;
  runId?: string;
  runKind?: RunKind;
  seq?: number;
}